+++ Runtime/AssemblyInfo.cs
@@ -1,3 +1,4 @@
 using System.Runtime.CompilerServices;
 
 [assembly: InternalsVisibleTo("com.sharq-it.sus.core.editor.tests")]
+[assembly: InternalsVisibleTo("com.sharq-it.sus.core.runtime.tests")]